+++ NEWS.md
@@ -13,6 +13,7 @@
 * Scripting: Renamed Tileset.collection to Tileset.isCollection (#3543)
 * Defold plugin: Allow overriding z value also when exporting to .collection (#3214)
 * Qt 6: Fixed invisible tileset tabs when only a single tileset is open
+* Qt 6: Fixed behavior of "Class of" selection popup
 * Fixed positioning of point object name labels (by Logan Higinbotham, #3400)
 * Fixed slight drift when zooming the map view in/out
 * Fixed compile against Qt 6.4